requirements.txt
asgiref==3.5.2
Django==4.0
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
uritemplate==4.1.1